# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/panel/preferences": {
      "get": {
        "operationId": "PanelScheduleController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Preferencias vigentes."
          },
          "401": {
            "description": "NO_SESSION."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Plantilla semanal y plazos configurables",
        "tags": [
          "panel-schedule"
        ]
      },
      "put": {
        "operationId": "PanelScheduleController_updatePreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Preferencias actualizadas."
          },
          "400": {
            "description": "INVALID_PREFERENCES o INVALID_RANGE."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar el plazo de confirmación, la antelación de serie y el horizonte de reserva",
        "tags": [
          "panel-schedule"
        ]
      }
    },
    "/api/panel/template": {
      "put": {
        "operationId": "PanelScheduleController_replaceTemplate",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Plantilla actualizada; las sesiones existentes no se tocan."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reemplazar la plantilla semanal completa",
        "tags": [
          "panel-schedule"
        ]
      }
    },
    "/api/panel/blocks/day": {
      "post": {
        "operationId": "PanelScheduleController_blockDay",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Día bloqueado."
          },
          "409": {
            "description": "SLOT_NOT_EMPTY."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bloquear un día completo (solo si no tiene sesiones activas)",
        "tags": [
          "panel-schedule"
        ]
      }
    },
    "/api/panel/blocks/day/{date}": {
      "delete": {
        "operationId": "PanelScheduleController_unblockDay",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Día desbloqueado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desbloquear un día (204 aunque no estuviera bloqueado)",
        "tags": [
          "panel-schedule"
        ]
      }
    },
    "/api/panel/blocks/slot": {
      "post": {
        "operationId": "PanelScheduleController_blockSlot",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Cupo bloqueado."
          },
          "409": {
            "description": "SLOT_NOT_EMPTY."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bloquear un cupo puntual (solo si está vacío)",
        "tags": [
          "panel-schedule"
        ]
      }
    },
    "/api/panel/blocks/slot/{date}/{time}": {
      "delete": {
        "operationId": "PanelScheduleController_unblockSlot",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "time",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cupo desbloqueado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desbloquear un cupo (204 aunque no estuviera bloqueado)",
        "tags": [
          "panel-schedule"
        ]
      }
    },
...
    "/api/panel/guardians": {
      "get": {
        "operationId": "PanelPeopleController_list",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de apoderados."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar apoderados (con conteo de niños y sesiones activas)",
        "tags": [
          "panel-people"
        ]
      },
      "post": {
        "operationId": "PanelPeopleController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Apoderado creado."
          },
          "409": {
            "description": "GUARDIAN_EXISTS."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear una ficha de apoderado",
        "tags": [
          "panel-people"
        ]
      }
    },
    "/api/panel/guardians/{id}": {
      "get": {
        "operationId": "PanelPeopleController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ficha del apoderado."
          },
          "404": {
            "description": "GUARDIAN_NOT_FOUND."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ficha de un apoderado: sus niños y el histórico de sesiones",
        "tags": [
          "panel-people"
        ]
      },
      "patch": {
        "operationId": "PanelPeopleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Apoderado actualizado."
          },
          "404": {
            "description": "GUARDIAN_NOT_FOUND."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar nombre o teléfono de un apoderado",
        "tags": [
          "panel-people"
        ]
      }
    },
    "/api/panel/guardians/{id}/children": {
      "post": {
        "operationId": "PanelPeopleController_createChild",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Niño creado."
          },
          "404": {
            "description": "GUARDIAN_NOT_FOUND."
          },
          "409": {
            "description": "CHILD_EXISTS."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Agregar un niño a la ficha de un apoderado (sin rango de edad)",
        "tags": [
          "panel-people"
        ]
      }
    },
    "/api/panel/children/{id}": {
      "patch": {
        "operationId": "PanelPeopleController_updateChild",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Niño actualizado."
          },
          "404": {
            "description": "CHILD_NOT_FOUND."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar nombre o edad de un niño",
        "tags": [
          "panel-people"
        ]
      }
    },
...
  },
  "info": {
    "title": "profesor-scheduling-api",
    "description": "Backend del agendamiento de la educadora diferencial.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}